intermediate revision - minor fix of the logic and dropped wrong "unzip" flag

+++ get-latest-chromedriver-json.sh
@@ -15,25 +15,34 @@ echo "NOTE: the following query will show just the  chromedriver download link:"
 jq -cr '.channels.Stable.downloads.chromedriver[]|select(.platform=="linux64").url' $JSONFILE
 DRIVERURL=$(jq -cr '.channels.Stable.downloads.chromedriver[]|select(.platform=="linux64").url' $JSONFILE)
 rm -f $JSONFILE
-echo "curl -k -I $DRIVERFILE $DRIVERURL"
-curl -k -I $DRIVERURL
-echo "curl -k -o $DRIVERFILE $DRIVERURL"
-curl -k -o $DRIVERFILE $DRIVERURL
+echo "curl -sk -I $DRIVERFILE $T"
+curl -sk -I $DRIVERURL
+echo "curl -sk -o $DRIVERFILE $DRIVERURL"
+curl -sk -o $DRIVERFILE $DRIVERURL
 echo "Verify contents of $DRIVERFILE"
 unzip -t $DRIVERFILE 
 # echo unzip -d '/tmp/' -u $DRIVERFILE
-echo 'Stopping chromedriver'
 # NOTE: prevent error with the lock:
 # cp: cannot create  regular file '...chromedriver': Text file busy
-killall chromedriver
-unzip -d '/tmp/' -f -u $DRIVERFILE
+ps ax | grep -q [c]hromedriver
+if [ ! $? ]
+then
+  echo 'Stopping chromedriver'
+  killall chromedriver
+fi
+# SCRIPTDIR=$(pwd)
+# cd $(dirname $DRIVERFILE)
+# ls $DRIVERFILE
+echo unzip -d '/tmp/' -u $DRIVERFILE
+unzip -d '/tmp/' -u $DRIVERFILE
 # TODO: handle
 # archive:  /tmp/chromedriver-linux64.zip
 # replace /tmp/chromedriver-linux64/LICENSE.chromedriver? [y]es, [n]o, [A]ll, [N]one, [r]ename: A
-
-cp /tmp/chromedriver-linux64/chromedriver ~/Downloads
-~/Downloads/chromedriver -version
+ls /tmp/chromedriver-linux64
+cp /tmp/chromedriver-linux64/chromedriver "$HOME/Downloads"
+"$HOME/Downloads/chromedriver" -version
 rm -f $DRIVERFILE
 # to check the chrome browser version on Linux need to run the executable with the version option:
 # google-chrome -version
 # Google Chrome 124.0.6367.60
+# cd $SCRIPTDIR
